validate channel admin integrity

+++ src/main/kotlin/platformapi/channelMemberList.kt
@@ -1,5 +1,9 @@
 package platformapi
 
+import dev.saibotma.persistence.postgres.jooq.enums.ChannelMemberRole
+import error.ApiException
+import error.managedChannelHasAdmin
+import error.resourceNotFound
 import io.ktor.application.*
 import io.ktor.http.*
 import io.ktor.request.*
@@ -19,6 +23,11 @@ suspend fun PipelineContext<Unit, ApplicationCall>.addMember(
     val member = call.receive<ChannelMemberWritePayload>()
     val channelId = location.channelDetails.channelId
     val result = database.transaction {
+        val channel = getChannel(channelId = channelId) ?: throw ApiException.resourceNotFound()
+        if (channel.isManaged!! && member.role == ChannelMemberRole.admin) {
+            throw ApiException.managedChannelHasAdmin()
+        }
+
         insertMember(
             member.toChannelMember(
                 channelId = channelId,
@@ -30,7 +39,6 @@ suspend fun PipelineContext<Unit, ApplicationCall>.addMember(
     call.respond(HttpStatusCode.Created, result)
 }
 
-// TODO(saibotma): Check that ther must be at least one admin when not managed
 // and that there must not be an admin when managed
 suspend fun PipelineContext<Unit, ApplicationCall>.updateMembers(
     location: ChannelList.ChannelDetails.ChannelMemberList,
@@ -39,12 +47,19 @@ suspend fun PipelineContext<Unit, ApplicationCall>.updateMembers(
     val channelId = location.channelDetails.channelId
     val members = call.receive<Array<ChannelMemberWritePayload>>()
     val result = database.transaction {
+        val channel = getChannel(channelId = channelId) ?: throw ApiException.resourceNotFound()
+        if (channel.isManaged!! && members.any { it.role == ChannelMemberRole.admin }) {
+            throw ApiException.managedChannelHasAdmin()
+        }
+
         val currentMembers = getMembersOf(channelId).map { it.toChannelMemberWrite() }
+
         val union = members.map { it.userId }.intersect(currentMembers.map { it.userId }).toList()
         val deleted = currentMembers.filter { !union.contains(it.userId) }.map { it.userId }
         val inserted = members.filter { member -> !currentMembers.map { it.userId }.contains(member.userId) }
             .map { it.toChannelMember(channelId, addedAt = now()) }
         val updated = members.filter { union.contains(it.userId) }
+
         deleteMembersOf(channelId, deleted)
         insertMembers(inserted)
         updateMembers(channelId = channelId, members = updated)
